# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
security:
  - bearerAuth: []
  - basicAuth: []
  - apiTokenAuth: []
...
paths:
...
  "/users/{user_id}/integration-tokens":
    get:
      tags:
        - Users
      description: lists integration bearer tokens for an API user.
      operationId: getUserIntegrationTokens
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: list of integration tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/IntegrationToken"
    post:
      tags:
        - Users
      description: creates a new integration bearer token for an API user.
      operationId: createUserIntegrationToken
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntegrationTokenRequest"
      responses:
        "200":
          description: created integration token
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/IntegrationToken"
                      - type: object
                        properties:
                          token:
                            type: string
  "/users/{user_id}/integration-tokens/{token_id}":
    delete:
      tags:
        - Users
      description: revokes an integration bearer token for an API user.
      operationId: deleteUserIntegrationToken
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
        - in: path
          name: token_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: token revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
    apiTokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Use the header format: `Authorization: token api_user:token`'
  schemas:
...
    IntegrationToken:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        last_used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateIntegrationTokenRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
...